servers - renamed JettyWebServer to Jetty7WebServer, same for module and properties
