+++ restx-server-testing/module.ivy
@@ -14,6 +14,7 @@
         <artifact type="jar"/>
     </publications>
     <dependencies>
+        <dependency org="io.restx" name="restx-server-jetty7" rev="latest.integration" conf="test->default" />
         <dependency org="io.restx" name="restx-server-jetty8" rev="latest.integration" conf="test->default" />
         <dependency org="io.restx" name="restx-server-simple" rev="latest.integration" conf="test->default" />
         <dependency org="io.restx" name="restx-server-tomcat" rev="latest.integration" conf="test->default" />
