+++ restx-server-testing/pom.xml
@@ -16,6 +16,11 @@
     <dependencies>
         <dependency>
             <groupId>io.restx</groupId>
+            <artifactId>restx-server-jetty7</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>io.restx</groupId>
             <artifactId>restx-server-jetty8</artifactId>
             <scope>test</scope>
         </dependency>
